Reorganize ETL modules into functional directories with backward compatibility
