+++ etl/geoprocess/__init__.py
@@ -1,0 +1,5 @@
+"""ETL geoprocessing package for spatial data processing."""
+
+from . import process
+
+__all__ = ['process']
