+++ etl/sde/__init__.py
@@ -1,0 +1,5 @@
+"""ETL SDE package for SDE database operations."""
+
+from . import load_sde
+
+__all__ = ['load_sde']
