+++ etl/staging/__init__.py
@@ -1,0 +1,5 @@
+"""ETL staging package for preparing data for processing."""
+
+from . import stage_files
+
+__all__ = ['stage_files']
