+++ etl/staging/stage_files.py
@@ -9,10 +9,10 @@ import shutil
 import zipfile
 from pathlib import Path
 
-from .sr_utils import SWEREF99_TM, WGS84_DD, detect_sr_from_geojson, validate_coordinates_magnitude
+from etl.utils.sr_utils import SWEREF99_TM, WGS84_DD, detect_sr_from_geojson, validate_coordinates_magnitude
 
 # Lazy ArcPy usage: import inside functions to avoid heavy init before logging
-from .utils import make_arcpy_safe_name
+from etl.utils.utils import make_arcpy_safe_name
 
 
 def _flatten_coordinates(coords):
